+++ R/bind_rows.R
@@ -2,7 +2,6 @@
 # Win-Vector LLC currently distributes this code without intellectual property indemnification, warranty, claim of fitness of purpose, or any other guarantee under a GPL3 license.
 
 
-
 #' Union two tables.
 #'
 #' Spark 2* union_all has issues ( https://github.com/WinVector/replyr/blob/master/issues/UnionIssue.md ),
@@ -179,6 +178,79 @@ r_replyr_bind_rows <- function(lst,
   res
 }
 
+#' Bind rows by a query. Assumes all tables structured identically.
+#'
+#' @param tableNames names of tables to concatinate (not empty)
+#' @param colNames names of columns
+#' @param my_db connection to where tables live
+#' @param tempNameGenerator temp name generator produced by replyr::makeTempNameGenerator, used to record dplyr::compute() effects.
+#' @param ... force later arguments to bind by name
+#' @param origTableColumn character, column to put original table name in.
+#'
+#' @examples
+#'
+#' my_db <- DBI::dbConnect(RSQLite::SQLite(), ":memory:")
+#' tempNameGenerator = makeTempNameGenerator("bind_rowsQ_demo")
+#' d1 <- replyr_copy_to(my_db, data.frame(x=1:2, y= 10:11), 'd1')
+#' d2 <- replyr_copy_to(my_db, data.frame(x=3:4, y= 13:14), 'd2')
+#' bind_rowsQ(c('d1', 'd2'), c('x', 'y'), my_db,
+#'            tempNameGenerator = tempNameGenerator,
+#'            origTableColumn = 'orig_table')
+#'
+#' @export
+#'
+bind_rowsQ <- function(tableNames, colNames, my_db,
+                       tempNameGenerator = makeTempNameGenerator("bind_rowsQ"),
+                       ...,
+                       origTableColumn = NULL) {
+  if(length(list(...))>0) {
+    stop("replyr::bind_rowsQ unexpected arguments")
+  }
+  nm <- tempNameGenerator()
+  qSel <- vapply(colNames,
+                 function(ci) {
+                   DBI::dbQuoteIdentifier(my_db, ci)
+                 }, character(1))
+  selTerms <- paste(qSel, collapse = ", ")
+  if(!is.null(origTableColumn)) {
+    selTerms <- paste0(selTerms,
+                       ", ",
+                       DBI::dbQuoteString(my_db, tableNames[[1]]),
+                       " ",
+                       DBI::dbQuoteIdentifier(my_db, origTableColumn))
+  }
+  qc <- paste0("CREATE TEMPORARY TABLE ",
+               DBI::dbQuoteIdentifier(my_db, nm),
+               " AS SELECT ",
+               selTerms,
+               " FROM ",
+               tableNames[[1]])
+  tryCatch(
+    r <- DBI::dbGetQuery(my_db, qc),
+    warning = function(w) { NULL })
+  if(length(tableNames)>1) {
+    for(i in 2:length(tableNames)) {
+      selTerms <- paste(qSel, collapse = ", ")
+      if(!is.null(origTableColumn)) {
+        selTerms <- paste0(selTerms,
+                           ", ",
+                           DBI::dbQuoteString(my_db, tableNames[[i]]),
+                           " ",
+                           DBI::dbQuoteIdentifier(my_db, origTableColumn))
+      }
+      qi <- paste0("INSERT INTO ",
+                   DBI::dbQuoteIdentifier(my_db,  nm),
+                   " SELECT ",
+                   selTerms,
+                   " FROM ",
+                   tableNames[[i]])
+      tryCatch(
+        r <- DBI::dbGetQuery(my_db, qi),
+        warning = function(w) { NULL })
+    }
+  }
+  dplyr::tbl(my_db, nm)
+}
 
 #' Bind a list of items by rows (can't use dplyr::bind_rows or dplyr::combine on remote sources).  Columns are intersected.
 #'
@@ -189,20 +261,28 @@ r_replyr_bind_rows <- function(lst,
 #' @param ... force other arguments to be used by name
 #' @param useDplyrLocal logical if TRUE use dplyr for local data.
 #' @param useSparkRbind logical if TRUE try to use rbind on Sparklyr data
+#' @param useQBind logical if TRUE try to use query based binding
 #' @param eagerTempRemoval logical if TRUE remove temps early.
 #' @param tempNameGenerator temp name generator produced by replyr::makeTempNameGenerator, used to record dplyr::compute() effects.
 #' @return single data item
 #'
 #' @examples
 #'
-#' d <- data.frame(x=1:2)
-#' replyr_bind_rows(list(d,d,d,d,d), useDplyrLocal= FALSE)
+#'
+#' my_db <- DBI::dbConnect(RSQLite::SQLite(), ":memory:")
+#' # my_db <- sparklyr::spark_connect(master = "local")
+#' d <- replyr_copy_to(my_db, data.frame(x = 1:2), 'd',
+#'                     temporary = TRUE)
+#' # dplyr::bind_rows(list(d, d))
+#' # # Argument 1 must be a data frame or a named atomic vector, not a tbl_dbi/tbl_sql/tbl_lazy/tbl
+#' replyr_bind_rows(list(d, d))
 #'
 #' @export
 replyr_bind_rows <- function(lst,
                              ...,
                              useDplyrLocal= TRUE,
                              useSparkRbind= TRUE,
+                             useQBind= TRUE,
                              eagerTempRemoval= FALSE,
                              tempNameGenerator= makeTempNameGenerator("replyr_bind_rows")) {
   if(length(list(...))>0) {
@@ -238,6 +318,30 @@ replyr_bind_rows <- function(lst,
       return(do.call(rbind, lst))
     }
   }
+  if(useQBind) {
+    # assuming all tables on same source
+    src <- dplyr_src_to_db_handle(replyr_get_src(lst[[1]]))
+    colnames <- colnames(lst[[1]])
+    tableNames <- lapply(lst,
+                         function(ti) {
+                           nm = tempNameGenerator()
+                           if(replyr_is_local_data(ti)) {
+                             ti %.>%
+                               seplyr::select_se(., colnames) %.>%
+                               replyr_copy_to(src, ., temporary = TRUE,
+                                              name = nm)
+                           } else {
+                             ti %.>%
+                               seplyr::select_se(., colnames) %.>%
+                               dplyr::compute(.,
+                                              name = nm)
+                           }
+                           nm
+                         })
+    return(bind_rowsQ(tableNames, colnames, src,
+                      tempNameGenerator = tempNameGenerator))
+  }
+  # nasty recursive fall-back
   r_replyr_bind_rows(lst, eagerTempRemoval, TRUE,
                      makeTempNameGenerator("bind_rows_priv"),
                      tempNameGenerator)